add logs in controller layer

+++ src/main/java/com/social_media/service/impl/FriendRequestServiceImpl.java
@@ -1,6 +1,5 @@
 package com.social_media.service.impl;
 
-import com.social_media.annotation.ValidateUserNotBlocked;
 import com.social_media.entity.FriendRequest;
 import com.social_media.entity.User;
 import com.social_media.exception.RequestNotAllowedException;
@@ -26,11 +25,9 @@ public class FriendRequestServiceImpl implements FriendRequestService {
 
     private static final String FRIEND_REQUEST_ALREADY_SENT_MESSAGE = "you have already sent a friend request";
 
-    private static final String FRIEND_REQUEST_PENDING_MESSAGE = "cannot accept friend request of the target user";
+    private static final String UNABLE_TO_UPDATE_DELETE_FRIEND_REQUEST_MESSAGE = "cannot update or delete friendRequest";
 
     private static final String FRIEND_REQUEST_NOT_FOUND_MESSAGE = "friend request not found";
-
-    private static final String UNABLE_TO_DELETE_FRIEND_REQUEST = "cannot delete friend request";
 
     private final FriendRequestRepository friendRequestRepository;
 
@@ -65,7 +62,6 @@ public class FriendRequestServiceImpl implements FriendRequestService {
 
     @Override
     @Transactional
-    @ValidateUserNotBlocked
     public FriendRequest updateFriendRequestStatus(User user, UUID requestId, FriendRequest.Status status) {
         log.info("updating friendRequest with id {}", requestId);
 
@@ -73,10 +69,13 @@ public class FriendRequestServiceImpl implements FriendRequestService {
                 .orElseThrow(() -> new ResourceNotFoundException(FRIEND_REQUEST_NOT_FOUND_MESSAGE));
 
         if (
+            FriendRequest.Status.BLOCKED.equals(friendRequest.getStatus()) ||
+            (
                 !user.getId().equals(friendRequest.getTargetUser().getId()) &&
                 FriendRequest.Status.ACCEPTED.equals(status)
+            )
         ) {
-            throw new RequestNotAllowedException(FRIEND_REQUEST_PENDING_MESSAGE);
+            throw new RequestNotAllowedException(UNABLE_TO_UPDATE_DELETE_FRIEND_REQUEST_MESSAGE);
         }
 
         friendRequest.setStatus(status);
@@ -90,7 +89,6 @@ public class FriendRequestServiceImpl implements FriendRequestService {
 
     @Override
     @Transactional
-    @ValidateUserNotBlocked
     public void deleteFriendRequest(User user, UUID requestId) {
         log.info("deleting friendRequest with id {}", requestId);
 
@@ -101,10 +99,13 @@ public class FriendRequestServiceImpl implements FriendRequestService {
         UUID targetUserId = friendRequest.getTargetUser().getId();
 
         if (
+            FriendRequest.Status.BLOCKED.equals(friendRequest.getStatus()) ||
+            (
                 !friendRequest.getUser().getId().equals(userId) &&
                 !targetUserId.equals(userId)
+            )
         ) {
-            throw new RequestNotAllowedException(UNABLE_TO_DELETE_FRIEND_REQUEST);
+            throw new RequestNotAllowedException(UNABLE_TO_UPDATE_DELETE_FRIEND_REQUEST_MESSAGE);
         }
 
         friendRequestRepository.delete(friendRequest);
@@ -113,7 +114,7 @@ public class FriendRequestServiceImpl implements FriendRequestService {
     }
 
     @Override
-    public Page<FriendRequest> getFriendRequestsByStatus(User user, FriendRequest.Status status, Pageable pageable) {
+    public Page<FriendRequest> getFriendRequestsByUserStatus(User user, FriendRequest.Status status, Pageable pageable) {
         UUID id = user.getId();
 
         log.info("fetching friendRequests of user with user {} and status {}", id, status);
